Remove `dataset` from `members.mjs` items — add `mActiveItem` and `mItems` module variables (#574)
* Initial plan

* Remove dataset from members.mjs items; add mActiveItem module variable with getter


* 20260311 @Mookse
- base changes to incorporate removal of datasets

* Add mItems array, items/registerItem getters, mRefreshCollection; pre-create popups in bots.mjs


* 20260312 @Mookse
- backend itemForms
- added `primaryCollectionTypes` array to team

* 20260312 @Mookse wip
- collections migration to `ciollections.mjs`
- removed bot-bar `mCreateBotThumb`

* 20260313 @Mookse wip
- small collection list fixes

* 20260313 @Mookse wip
- popup improvements

* 20260313 @Mookse wip
- manage active item clicks

* 20260314 @Mookse
- stable fix with deeper errors on
- obscure
- relive
- summarize
- files

* 20260314 @Mookse
- minor update on relive

---------
