+++ inc/json-schemas/intelligences/avatar-intelligence-1.2.json
@@ -35,6 +35,10 @@
         }
       ]
     },
+    "itemForms": [
+      "chat",
+      "file"
+    ],
     "name": "instructions-personal-avatar-bot",
     "purpose": "Member Avatar is the Primary Controller for a MyLife member's bot legion",
     "retirable": false,
