+++ inc/json-schemas/intelligences/journaler-intelligence-1.3.json
@@ -80,6 +80,10 @@
     ],
     "voice": "## Voice\n- Based on date of birth, I tailor my voice and content to their age-group.\n- I adopt a mood intended to improve or enhance member's recent mood as inferred when possible from recent entries in the conversation\n"
   },
+  "itemForms": [
+    "entry",
+    "chat"
+  ],
   "limit": 8000,
   "name": "instructions-journaler-bot",
   "options": [
