+++ views/assets/css/bots.css
@@ -356,6 +356,11 @@
     min-height: 2.0rem;
     width: 100%;
 }
+.collection-bot {
+    display: flex;
+    height: 1.6rem;
+    width: 1.6rem;
+}
 .collection-icon {
     display: flex;
     flex: 0 0 auto;
@@ -422,18 +427,34 @@
     overflow-x: hidden;
     overflow-y: auto;
 }
-.collection-popup.popup-container {
-    display: flex;
+.collection-popup {
+    background-color: aliceblue;
+    display: none;
     flex: 1 0 auto;
     flex-direction: column;
     height: fit-content;
     margin: 0;
+    max-height: 80vh;
+    max-width: 80vw;
+    left: 10vw;
+    opacity: 0;
     overflow: auto;
-	opacity: 0; /* Start with hidden popup */
-    transition: opacity 0.3s; /* Smooth transition */
     padding: 0;
-    width: 40em;
-    z-index: var(--collection-item-popup-z-index); /* above sidebar and other popups */
+    position: absolute;
+    pointer-events: none; /* Allow interaction when visible */
+    top: 10vh;
+    transition: opacity 0.5s ease; /* Smooth transition */
+    width: auto;
+    z-index: 1; /* default to creation level */
+
+}
+.collection-popup.popup-active {
+    z-index: 10;
+}
+.collection-popup.show {
+    display: flex;
+    opacity: 1;
+    pointer-events: auto; 
 }
 .collection-popup-body {
     display: flex;
