+++ views/assets/html/_bots.html
@@ -44,52 +44,4 @@
     <div id="team-name" class="team-name">MyLife Team</div>
     <span id="add-team-member-icon" class="fas fa-person-circle-plus add-team-member-icon" name="add-team-member"></span>
 </div>
-<div class="collections-container" id="collections-container">
-    <div class="collections-titlebar" id="collections-titlebar" name="collections-titlebar">
-        <div class="collections-spacer" id="collections-titlebar-spacer"></div>
-        <div class="collections-title" id="collections-title" name="collections-title">Scrapbook</div>
-        <div class="bot-options-dropdown" id="collections-options-dropdown"></div><!-- drop-down caret to trigger slide-down of container below -->
-    </div>
-    <div class=" bot-options collections-options hidden" id="collections-options" name="collections-options">
-        <!-- Bot Name -->
-        <div class="collections-description show" id="collections-description">
-            Click on a category to browse entries
-        </div>
-        <div id="collections-collections" class="collections"><!-- from here _could_ be dynamic, not just listing -->
-            <div id="collection-memory" class="collection memory-collection">
-                <div id="collection-bar-memory" class="collection-bar">
-                    <div id="collection-icon-memory" class="collection-icon collection-icon memory-icon"></div>
-                    <div id="collection-title-memory" class="collection-title collection-title memory-title">Memories</div>
-                    <span class="fa-solid fa-refresh collection-refresh memory-refresh" id="collection-refresh-memory"></span>
-                </div>
-                <div id="collection-list-memory" class="collection-list memory-list">
-                    <!-- memory items populate here -->
-                    <div class="collection-item memory-item">None</div>
-                </div>
-            </div>
-            <div id="collection-entry" class="collection entry-collection">
-                <div id="collection-bar-entry" class="collection-bar">
-                    <div id="collection-icon-entry" class="collection-icon collection-icon entry-icon"></div>
-                    <div id="collection-title-entry" class="collection-title collection-title entry-title">Entries</div>
-                    <span class="fa-solid fa-refresh collection-refresh entry-refresh" id="collection-refresh-entry"></span>
-                </div>
-                <div id="collection-list-entry" class="collection-list entry-list">
-                    <!-- entry items populate here -->
-                    <div class="collection-item entry-item">None</div>
-                </div>
-            </div>
-            <div id="collection-file" class="collection file-collection">
-                <div id="collection-bar-file" class="collection-bar">
-                    <div id="collection-icon-file" class="collection-icon collection-icon"></div>
-                    <div id="collection-title-file" class="collection-title collection-title">Files</div>
-                    <span class="fa-solid fa-refresh collection-refresh" id="collection-refresh-file"></span>
-                </div>
-                <div id="collection-list-file" class="collection-list">
-                    <!-- file items populate here -->
-                    <div class="collection-item file-item">None</div>
-                </div>
-            </div>
-        </div>
-        <button id="collections-upload" class="bot-options-button button upload collections-upload" name="bot-upload">Upload</button>
-    </div>
-</div>
+<div class="collections-container" id="collections-container"></div>
