+++ views/assets/js/globals.mjs
@@ -36,6 +36,7 @@ let mActiveHelpType, // active help type, currently entire HTMLDivElement
     mLoaded = false,
     mLogoutButton,
     mMainContent,
+    mOverlays,
     mNavigation,
     mNavigationHamburger,
     mNavigationHelp,
@@ -783,6 +784,7 @@ class Globals {
             mNavigationHelp = document.getElementById('navigation-help')
             mNavigationHelpIcon = document.getElementById('navigation-help-icon')
             mNavigationMenu = document.getElementById('navigation-menu')
+            mOverlays = document.getElementById('overlays')
             mPage = document.getElementById('page-header')
             mSidebar = document.getElementById('sidebar')
                 ?? document.getElementById('bot-container')
@@ -909,19 +911,6 @@ class Globals {
 		}
 		a.length = 0
 	}
-    /**
-     * Operates on a dataset to clear all frontend-defined keys.
-     * @param {DOMStringMap} dataset - The dataset to clear
-     * @returns {void}
-     */
-    clearDataset(dataset){
-        if(!(dataset instanceof DOMStringMap))
-            return
-        for(let key in dataset){
-            if(dataset.hasOwnProperty(key))
-                delete dataset[key]
-        }
-    }
     /**
      * Clears an element of its contents, brute force currently via innerHTML.
      * @param {HTMLElement} element - The element to clear.
@@ -1130,12 +1119,36 @@ class Globals {
         }
     }
     /**
+     * Determines whether an element is hidden via class or inline styles.
+     * @param {HTMLElement} element - The element to check for visibility
+     * @returns {boolean} - Whether the element is hidden
+     */
+    isHidden(element){
+        const { classList, } = element
+        const { display, visibility, } = getComputedStyle(element)
+        const hidden = element.classList.contains('hide')
+            || element.classList.contains('hidden')
+            || element.classList.contains('fade-out')
+            || (display ?? 'none') === 'none'
+            || (visibility ?? 'hidden') === 'hidden'
+        return hidden
+    }
+    /**
      * Determines whether the bot is a `proxy agent` given `type`.
      * @param {string} type - The type to check
      * @returns {boolean} - Whether the bot is a `proxy agent`
      */
     isProxy(type){
         return type=='proxy'
+    }
+    /**
+     * Pluralizes the last word in a string.
+     * @param {string} phrase - The content to pluralize
+     * @param {boolean} allCaps - Whether to capitalize all words in the phrase, defaults to `false` (only capitalizes first word)
+     * @returns {string} - The pluralized content
+     */
+    pluralize(phrase, allCaps=false){
+        return mPluralize(phrase, allCaps)
     }
     /**
      * Remove an element from the DOM based upon its class name of `input-disappear`.
@@ -1264,6 +1277,9 @@ class Globals {
     }
     get newGuid(){ 
         return mNewGuid()
+    }
+    get overlays(){
+        return mOverlays
     }
     get page(){
         return mPage
@@ -1671,6 +1687,39 @@ async function mLogout(){
         window.location.href = '/'
     else
         console.error('mLogout::failure', response)
+}
+/**
+ * Pluralizes the last word in a string.
+ * @param {string} phrase - The content to pluralize
+ * @param {boolean} allCaps - Whether to capitalize all words in the phrase
+ * @returns {string} - The pluralized content
+ */
+function mPluralize(phrase, allCaps){
+    if(typeof phrase !== 'string')
+        return phrase
+    phrase = phrase.trim()
+    const parts = phrase.split(/\s+/)
+    let lastWord = parts.pop()
+    // Basic pluralization rules
+    if(lastWord.endsWith('y') && !/[aeiou]y$/i.test(lastWord))
+        lastWord = lastWord.slice(0, -1) + 'ies'
+    else if (/(s|sh|ch|x|z)$/i.test(lastWord))
+        lastWord = lastWord + 'es'
+    else
+        lastWord = lastWord + 's'
+    parts.push(lastWord)
+    const response = parts
+        .map((part, index)=>{
+            if(allCaps || index===0)
+                part = capitalize(part)
+            return part
+        })
+        .join(' ')
+    return response
+    function capitalize(word){
+        word = word.trim()
+        return word.charAt(0).toUpperCase() + word.slice(1)
+    }
 }
 /**
  * Scrolls overflow of passed element to bottom.
