+++ views/layout.html
@@ -64,9 +64,7 @@
 			</div>
 		</div>
 	</div>
-    <!-- Page Footer -->
-    <footer>
-		<p>&copy; 2025 MyLife. All rights reserved.</p>
-	</footer>
+    <footer id="#footer"><p>&copy; 2025 MyLife. All rights reserved.</p></footer>
+	<div id="overlays"></div>
 </body>
 </html>
