+++ views/members.html
@@ -4,15 +4,12 @@
 	</div>
 	<div id="chat-system" class="chat-section"></div>
 	<div id="chat-separator" class="chat-separator">
-		<div id="chat-active-items" class="chat-active-items">
-			<img id="chat-active-item-thumb" class="chat-active-item-thumb" alt="Active item thumbnail"></img>
-			<div id="chat-active-item" class="chat-active-item">
-				<div id="chat-active-item-icon" class="fas fa-square chat-active-item-icon"></div>
-				<div id="chat-active-item-status" class="chat-active-item-status"></div>
-				<button id="chat-active-item-button" class="button chat-active-item-button" title="chat-active-item-button" type="button" disabled></button>
-				<div id="chat-active-item-title" class="chat-active-item-title"></div>
-				<div id="chat-active-item-close" class="fas fa-close chat-active-item-close"></div>
-			</div>
+		<div id="chat-active-item" class="chat-active-item">
+			<div id="chat-active-item-icon" class="fas fa-square chat-active-item-icon"></div>
+			<div id="chat-active-item-status" class="chat-active-item-status"></div>
+			<button id="chat-active-item-button" class="button chat-active-item-button" title="chat-active-item-button" type="button" disabled></button>
+			<div id="chat-active-item-title" class="chat-active-item-title"></div>
+			<div id="chat-active-item-close" class="fas fa-close chat-active-item-close"></div>
 		</div>
 		<div id="audio-popup" class="audio-popup"></div>
 	</div>
